requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.8.0
loompy>=3.0.6

# Metabolic modeling and flux analysis
cobra>=0.26.0
optlang>=1.5.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine learning and statistics
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Jupyter and notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Data I/O
h5py>=3.1.0
tables>=3.6.0
openpyxl>=3.0.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
requests>=2.26.0

# Quality control
scrublet>=0.2.3

# Gene set analysis
gseapy>=0.10.0

# Optional: GPU acceleration
# cupy-cuda11x>=10.0.0  # Uncomment if CUDA is available
